requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
xgboost==1.7.6
imbalanced-learn==0.10.1
scapy==2.5.0
langchain-groq==0.0.3
langchain-core==0.0.10
plotly==5.17.0
python-dotenv==1.0.0
joblib==1.3.2
seaborn==0.12.2
matplotlib==3.7.2